.cqfdrc: remove build directory between builds for sfl_ci
This commit is a backport from kirkstone branch.

Original commit message:
The sfl_ci flavor builds multiple images with different distro which can
create conflicts in the workdir.

Also, some recipes can fail during build if part of the cache is
invalidated and a workdir is already existing.
For example, the libvirt recipe fails with a pseudo error during
do_install if a workdir was created with a previous build and do_install
should rerun. This is because the do_install task of libvirt triggers a
"setup.py install" command for a cython module which also triggers
the build command. However, building is not supposed to happen in
do_install and conflict with previous pseudo setup.

To prevent such errors, remove the build directory between each image
build.

Change-Id: Ie11d2d09a29ea074c78723185e22ae9b8a6d5e1a
Signed-off-by: Elinor Montmasson <elinor.montmasson@savoirfairelinux.com>